Use ImplicitOutline for appendChildren case
Not sure if this is what this flag is meant to indicate, but it passes
two more toml-test cases: a top-level array section should not be
serialized as an object with type:array and value, but rather just a
simple array. Flagging it as ImplicitOutline, then serializing that case
accordingly ensures this.